+++ package/models/unet.py
@@ -100,4 +100,4 @@ class UNet(pl.LightningModule):
         return preds
 
     def configure_optimizers(self):
-        return torch.optim.Adam(self.parameters(), lr=self.lr)
+        return torch.optim.AdamW(self.parameters(), lr=self.lr, weight_decay=1e-4)