+++ package/datamodules/__init__.py
@@ -1,1 +1,1 @@
-from .sentinel2_60m_datamodule import Sentinel2_60mDataModule
+from .sentinel2_60m_datamodule import Sentinel2_60mDataModule, createSentinel2_60mDataModule