+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.2.0] - 2025-11-12
+
+### Changed
+
+- Moved DataModule and LightningModule creation out of the trainer script
+
 ## [0.1.0] - 2025-11-09
 
 ### Added